Adds event recording and reporting
Implements the ability to record and report events like potholes or speed bumps.

This commit introduces an event queue and data structure to store event information (type, GPS coordinates, timestamp). The TCP server task can now receive commands to record these events, and the SD card task writes them to the data log with the time of occurrence. Also, it adds event data to the JSON response to signal to the user that the event has been recorded successfully.